requirements.txt
python-dotenv
git+https://github.com/DisnakeDev/disnake#egg=disnake[voice]

docs/_build